+++ bigmler/anomaly/dispatcher.py
@@ -125,6 +125,7 @@ def compute_output(api, args):
     anomalies = None
     fields = None
     test_dataset = None
+    datasets = None
     # no multi-label support at present
 
     # variables from command-line options
@@ -158,17 +159,34 @@ def compute_output(api, args):
         log = args.log_file
         # If --clear_logs the log files are cleared
         clear_log_files([log])
-
-    source, resume, csv_properties, fields = ps.source_processing(
-        api, args, resume,
-        csv_properties=csv_properties,
-        session_file=session_file, path=path, log=log)
-
-    datasets, resume, csv_properties, fields = pd.dataset_processing(
-        source, api, args, resume,
-        fields=fields,
-        csv_properties=csv_properties,
-        session_file=session_file, path=path, log=log)
+    if args.source_file:
+        # source is retrieved from the contents of the given local JSON file
+        source, csv_properties, fields = u.read_local_resource(
+            args.source_file,
+            csv_properties=csv_properties)
+    else:
+        # source is retrieved from the remote object
+        source, resume, csv_properties, fields = ps.source_processing(
+            api, args, resume,
+            csv_properties=csv_properties,
+            session_file=session_file, path=path, log=log)
+    if args.dataset_file:
+        # dataset is retrieved from the contents of the given local JSON file
+        model_dataset, csv_properties, fields = u.read_local_resource(
+            args.dataset_file,
+            csv_properties=csv_properties)
+        if not args.datasets:
+            datasets = [model_dataset]
+            dataset = model_dataset
+        else:
+            datasets = u.read_datasets(args.datasets)
+    if not datasets:
+        # dataset is retrieved from the remote object
+        datasets, resume, csv_properties, fields = pd.dataset_processing(
+            source, api, args, resume,
+            fields=fields,
+            csv_properties=csv_properties,
+            session_file=session_file, path=path, log=log)
     if datasets:
         dataset = datasets[0]
         if args.to_csv is not None:
@@ -198,10 +216,18 @@ def compute_output(api, args):
             dataset, api, args, resume, fields=fields,
             session_file=session_file, path=path, log=log)
         datasets[0] = dataset
-
-    anomalies, anomaly_ids, resume = pa.anomalies_processing(
-        datasets, anomalies, anomaly_ids, api, args, resume, fields=fields,
-        session_file=session_file, path=path, log=log)
+    if args.anomaly_file:
+        # anomaly is retrieved from the contents of the given local JSON file
+        anomaly, csv_properties, fields = u.read_local_resource(
+            args.anomaly_file,
+            csv_properties=csv_properties)
+        anomalies = [anomaly]
+        anomaly_ids = [anomaly['resource']]
+    else:
+        # anomaly is retrieved from the remote object
+        anomalies, anomaly_ids, resume = pa.anomalies_processing(
+            datasets, anomalies, anomaly_ids, api, args, resume, fields=fields,
+            session_file=session_file, path=path, log=log)
     if anomalies:
         anomaly = anomalies[0]